Some small changes.
Raise ImportError instead of SystemExit when $DISPLAY is not set, so
regrtest.py will do the right thing.

Add a call to gl.clear() to clear the window before drawing in it.

Add some verbose prints for completeness.

Use gl.v2i() for the coordinates of one of the lines, for a little bit
of variety.

Reduce the time the window is displayed from 5 to 2 seconds.

--HG--
branch : legacy-trunk
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/python/branches/legacy-trunk%407363