A couple of grammatical nits.
Re-sequenced the function descriptions so that the formatting is described
before the assumption is made that the reader has a clue about the formatting.

Moved examples to be closer to the functions being demonstrated.

Added example of saferepr() result.

--HG--
branch : legacy-trunk
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/python/branches/legacy-trunk%408373