Preliminary LaTeX2HTML 98.1p1 support.  Includes patch from Ross Moore to get some funky quoting of ']' in \item[...] to work right without having to do really ugly things to the documents themselves.
There are a lot of things relating to indexing that are commented out;
parts need to be examined and dealt with with respect to changes in
LaTeX2HTML internals.  I'll work with Ross to see what it takes to make
this sort of stuff reasonable.

--HG--
branch : legacy-trunk
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/python/branches/legacy-trunk%4010018