do_cmd_seetext():  Clean up so we get proper nesting of <div>s.  This 		   fixes a rendering problem on IE5.
General adjustments to the table* environments, including using
<thead> and <tbody>.  Attempt to adjust the vertical alignment of the
table cells so that the baseline of the first cell matches the
baseline of the remaining cells:  When the first cell is small and the
second cell of the same row is multi-line, the first cell was
vertically centered by default.  Specifying valign=baseline fixes the
problem on IE, but Netscape seems to ignore both valign=top and
valign=baseline (even though valign is NS's fault!).  Make the
horizontal alignment of 'p' columns left instead of center (for the
headings).

--HG--
branch : legacy-trunk
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/python/branches/legacy-trunk%4013272