Use the \setshortversion macro to specify the abbreviated version number. This partially addresses SourceForge bug #114318.
Fix the release number -- this is supposed to be the *Python* release,
not the documentation release!  The later is given by the combination of
the Python release number and the date.  Add a comment so I can keep it
straight.

--HG--
branch : legacy-trunk
extra : convert_revision : svn%3A6015fed2-1504-0410-9fe1-9d1591cc4771/python/branches/legacy-trunk%4017413